# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Central management server for the OpenLabStats fleet. Manages agent enrollment, software mappings, lab groupings, and reporting.",
        "title": "OpenLabStats Server API",
        "contact": {},
        "version": "0.1.3"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/settings": {
            "get": {
                "description": "Returns global configuration for agents and server.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get system settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.SystemSettings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates global configuration for agents and server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update system settings",
                "parameters": [
                    {
                        "description": "Settings payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.SystemSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "store.SystemSettings": {
            "type": "object",
            "properties": {
                "autoUpdateEnabled": {
                    "description": "Staggered auto-update rollout. AutoUpdateEnabled is the master switch;\nRolloutMaxConcurrent bounds how many agents may be installing at once\n(0 = unlimited); RolloutGraceSeconds is how long a slot is held before an\nunfinished update is assumed stuck and the slot is freed;\nTargetAgentVersion pins the rollout target (\"\" = auto-track the newest\ninstaller on disk, per platform).",
                    "type": "boolean"
                },
                "heartbeatIntervalSeconds": {
                    "type": "integer"
                },
                "maintenanceWindowEnd": {
                    "description": "HH:mm",
                    "type": "string"
                },
                "maintenanceWindowStart": {
                    "description": "HH:mm",
                    "type": "string"
                },
                "minAgentVersion": {
                    "type": "string"
                },
                "rolloutGraceSeconds": {
                    "type": "integer"
                },
                "rolloutMaxConcurrent": {
                    "type": "integer"
                },
                "staleTimeoutDays": {
                    "type": "integer"
                },
                "targetAgentVersion": {
                    "type": "string"
                },
                "updateIntervalSeconds": {
                    "type": "integer"
                }
            }
        },
...
    }
}